refactor: 인터페이스 이동 및 사용되지 않는 인터페이스 제거
- FollowListDto 인터페이스를 follow.ts로 이동하여 파일 구조 정리
- 사용되지 않는 CommentListDto 인터페이스 제거
