+++ src/interfaces/user/userInfo.ts
@@ -26,19 +26,3 @@ export interface UserProfileDto extends UserPwDto {
   profilePicture: string;
   statusMessage: string;
 }
-
-export interface FollowListDto {
-  userId?: string;
-  email: string;
-  pageSize: number;
-  page: number;
-}
-
-export interface CommentListDto {
-  userId?: string;
-  email?: string;
-  sort: string;
-  pageSize: number;
-  cursor: string;
-  isBefore: boolean;
-}
